+++ group30_project/src/main/java/group30/Subtraction.java
@@ -2,8 +2,8 @@ package group30;
 
 public class Subtraction {
 
-    float calculate(float x, float y) {
-        float result = x - y;
+    double calculate(double x, double y) {
+        double result = x - y;
         return result;
     }
 }
